refactor(sources): deduplicate content utilities between issues and discussions
Extract hasCodeBlock, COMMENT_NOISE_RE, and truncateBody into
github-common.ts. These were identical implementations in both
issues.ts and discussions.ts.

The shared COMMENT_NOISE_RE uses the discussions variant which
also filters 'thanks/thank you' noise comments.
